tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/changing_keyframes_before.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - changing keyframes before animation start</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="The values used for the keyframes are
   snapshotted at the time the animation starts (and not before)">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            left: 0px;
         }
         to {
            left: -100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .test {
         animation: slide 1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationend', function() {
            clearTimeout(timeout);
            var passed = getComputedStyle(div, null).left == '200px';
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         });
         var toFrame = document.styleSheets[0].cssRules[0].cssRules[1];
         toFrame.style.left = '200px';
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>Below there should be a blue square sliding right for 1 second.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
